# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Router API
  version: 0.10.0
  summary: Single-user LLM gateway with named inference keys, provider configuration, fallback groups and usage metrics.
  description: |
    Auto Router is a single-user model gateway with provider-level upstream credentials, named client
    inference keys, and ordered fallback groups (simple, medium and complex). The first owner password
    is set once through the local setup endpoint; later access uses a password-backed browser session.

    Management APIs use the `auto_router_admin_session` cookie. Inference calls use a named inference
    key as `Authorization: Bearer <key>`. Provider secrets are write-only. Groups reference enabled
    provider/model pairs, and each group's array order is the fallback order.

    OpenAI-compatible ingress is available at `/v1/chat/completions` and `/v1/responses`; native
    Anthropic ingress is available at `/v1/messages`, and native Gemini ingress at
    `/v1beta/models/{model}:generateContent`. Native Anthropic and Gemini ingress requires the
    matching upstream protocol. OpenAI Chat text-only requests can be translated to Anthropic or
    Gemini for a limited supported subset; unsupported fields are rejected before upstream delivery.

    `GET /admin/v1/logs/summary` reports final client request success rate, token observations and
    60-second completed upstream output-token TPS. `GET /admin/v1/logs/stats` can aggregate actual
...
  license:
    name: MIT
...
servers:
...
security: []
paths:
...
  /admin/v1/setup/status:
    get:
      tags: [Setup]
      summary: Check whether the owner password has been set
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Setup status.
          content:
            application/json:
              schema:
                type: object
                required: [password_set]
                properties:
                  password_set: {type: boolean}
        '503':
          $ref: '#/components/responses/AdminError'
  /admin/v1/setup/password:
    post:
      tags: [Setup]
      summary: Set the initial owner password
      operationId: setInitialPassword
      description: Available only before a password exists and only from loopback with a same-origin request. Password length is 12–1024 bytes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string, minLength: 12, maxLength: 1024, writeOnly: true}
      responses:
        '201':
          description: Initial password set.
          content:
            application/json:
              schema:
                type: object
                required: [password_set]
                properties:
                  password_set: {type: boolean, const: true}
        '400': {$ref: '#/components/responses/AdminError'}
        '403': {$ref: '#/components/responses/AdminError'}
        '409': {$ref: '#/components/responses/AdminError'}
        '429': {$ref: '#/components/responses/AdminError'}
        '503': {$ref: '#/components/responses/AdminError'}
...
  /admin/v1/session:
    post:
      tags: [Admin]
      summary: Log in with the owner password
      operationId: createAdminSession
      description: Checks the single owner's password and sets an in-memory HttpOnly, SameSite=Strict browser session cookie. Login is same-origin and rate limited.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string, writeOnly: true}
      responses:
        '200':
          description: Session created.
          headers:
            Set-Cookie:
              description: "The `auto_router_admin_session` cookie, scoped to `/admin`."
              schema: {type: string}
          content:
            application/json:
              schema:
                type: object
                required: [expires_at]
                properties:
                  expires_at: {type: string, format: date-time}
        '400': {$ref: '#/components/responses/AdminError'}
        '401': {$ref: '#/components/responses/AdminError'}
        '403': {$ref: '#/components/responses/AdminError'}
        '413': {$ref: '#/components/responses/AdminError'}
        '429': {$ref: '#/components/responses/AdminError'}
    get:
      tags: [Admin]
      summary: Probe the current session
      operationId: getAdminSession
      description: |
        Reports the single authenticated owner identity. A `401` means the browser should show the login form.
      security:
        - adminSession: []
      responses:
        '200':
          description: The authenticated identity.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionProbe'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientScope'
    delete:
      tags: [Admin]
      summary: End the current session
      operationId: deleteAdminSession
      description: |
        Destroys the session the request was authenticated with and clears the cookie with
        `Max-Age=0` on the same `Path` it was issued with.
      security:
        - adminSession: []
      responses:
        '200':
          description: The session was ended.
          headers:
            Set-Cookie:
              description: The cleared session cookie.
              schema: {type: string}
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema:
                type: object
                required: [logged_out]
                properties:
                  logged_out: {type: boolean, const: true}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientScope'
...
  /admin/:
    get:
      tags: [Dashboard]
      summary: Serve the dashboard shell
      operationId: getDashboardPage
      description: |
        Serves the bundled Vue SPA shell (`text/html`). The same bytes are served for `/admin/`,
        `/admin/login`, `/admin/providers`, `/admin/pairs`,
        `/admin/settings` and `/admin/keys`:
        the shell is structurally independent of the database, and every number on a page is fetched
        by the browser from an authenticated `/admin/v1/*` endpoint. That is why these paths can be
        served without a credential and why they disclose nothing.

        An unknown path under `/admin/` is a `404`: this is an exact route lookup, not an SPA
        wildcard fallback.

        Registered only when the management surface is mounted. With `auth.enabled` or
        `admin.enabled` off, every `/admin/*` path answers `404` exactly as it did before the
        surface existed.
      security: []
      responses:
        '200':
          description: The page shell.
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema: {type: string}
            Content-Security-Policy:
              schema: {type: string}
          content:
            text/html:
              schema: {type: string}
        '404':
          description: Unknown page, or the management surface is not mounted.
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /admin/login:
    get:
      tags: [Dashboard]
      summary: Serve the login shell
      operationId: getDashboardLoginPage
      description: See `GET /admin/`. The shell reads `GET /admin/v1/session` and redirects to this path on `401`.
      security: []
      responses:
        '200':
          description: The page shell.
          content:
            text/html:
              schema: {type: string}
        '404':
          description: The management surface is not mounted.
  /admin/providers:
    get:
      tags: [Dashboard]
      summary: Serve the providers page shell
      operationId: getDashboardProvidersPage
      description: See `GET /admin/`.
      security: []
      responses:
        '200':
          description: The page shell.
          content:
            text/html:
              schema: {type: string}
        '404':
          description: The management surface is not mounted.
  /admin/pairs:
    get:
      tags: [Dashboard]
      summary: Serve the route bindings page shell
      operationId: getDashboardPairsPage
      description: See `GET /admin/`. This exact, data-free shell route supports GET and HEAD; binding data comes from `/admin/v1/pairs`.
      security: []
      responses:
        '200':
          description: The page shell.
          content:
            text/html:
              schema: {type: string}
        '404':
          description: The management surface is not mounted.
  /admin/settings:
    get:
      tags: [Dashboard]
      summary: Serve the settings page shell
      operationId: getDashboardSettingsPage
      description: See `GET /admin/`.
      security: []
      responses:
        '200':
          description: The page shell.
          content:
            text/html:
              schema: {type: string}
        '404':
          description: The management surface is not mounted.
  /admin/keys:
    get:
      tags: [Dashboard]
      summary: Serve the credentials page shell
      operationId: getDashboardKeysPage
      description: See `GET /admin/`. This exact, data-free shell route supports GET and HEAD; credential data comes from `/admin/v1/keys`.
      security: []
      responses:
        '200':
          description: The page shell.
          content:
            text/html:
              schema: {type: string}
        '404':
          description: The management surface is not mounted.
...
components:
  securitySchemes:
    inferenceKey:
      type: http
      scheme: bearer
      description: |
        A named inference key sent as `Authorization: Bearer <key>`.
    adminSession:
      type: apiKey
      in: cookie
      name: auto_router_admin_session
      description: |
        The browser session issued by `POST /admin/v1/session`. Accepted only under `/admin/`; it
        cannot authenticate inference requests.
  headers:
    XRequestId:
      description: |
        The correlation identifier for this request. An inbound value matching
        `^[A-Za-z0-9._-]{1,128}$` is echoed; anything else is replaced by a generated identifier. The
        same value is recorded in the process log and, for requests that reach the forwarding path,
        in the routing log.
      schema: {type: string}
...
  responses:
    Unauthorized:
      description: |
        No credential was presented, or the presented one was not accepted
        (`401 invalid_api_key`). The response carries `WWW-Authenticate: Bearer`.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OpenAIError'}
    InsufficientScope:
      description: The credential is valid but does not carry the scope this path requires (`403 insufficient_scope`).
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OpenAIError'}
...
    MethodNotAllowed:
      description: The path exists but not for this method. The mux answers with an `Allow` header.
      headers:
        Allow:
          schema: {type: string}
...
    AdminError:
      description: A management-surface failure, using the management envelope.
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/AdminError'}
...
  schemas:
...
    SessionProbe:
      type: object
      required: [authenticated, name, scopes, expires_at]
      properties:
        authenticated: {type: boolean, const: true}
        name: {type: string, description: The single owner identity name.}
        scopes:
          type: array
          items: {type: string, enum: [management]}
        expires_at:
          type: string
          description: The session's absolute expiry.